+++ config/web.php
@@ -102,7 +102,7 @@ $config = [
             ],
         ], // }}} 
         'mailer' => [ // {{{ 
-            'class' => 'yii\swiftmailer\Mailer',
+            'class' => 'yii\symfonymailer\Mailer',
             // send all mails to a file by default. You have to set
             // 'useFileTransport' to false and configure transport
             // for the mailer to send real emails.